signer: move spx keygen, sign and verify API
This was previously part of `securesystemslib.keys` interface.
In preparation for deprecation of this interface in favor of the
new signer API, spx-related functionality is moved.

This is a backwards incompatible change and a slight feature
degradation:

- default keyid is computed differently than before
- from_priv_key_uri (file, envvar) is no longer available

The advantage is that this is now completely independent of any
legacy securesystemslib code.

See usage example in docstring.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>